refactored upgraders
