+++ pkg/upgraders/osd/upgrader_test.go
@@ -1,4 +1,4 @@
-package osd_cluster_upgrader
+package osd
 
 import (
 	"fmt"
